+++ droidforge/moduleinvalid.h
@@ -10,7 +10,7 @@ public:
     QString name() const { return "invalid"; };
     QString faceplate() const { return "faceplate-blind-front"; };
     QString title() const { return "Invalid"; };
-    unsigned hp() const { return 1; };
+    float hp() const { return 1; };
 };
 
 #endif // MODULEINVALID_H
